feature/EFES-18 Смена пароля объединена в одно модальное окно с редактированием пользователя
