--- src/app/modules/profile/_modals/edit-personal-info/change-password/change-password.component.html
@@ -1,12 +1,0 @@
-<div class="tui-row jc-center" [formGroup]="passwordForm">
-  <div class="tui-col_md-6">
-    <div class="tui-space_bottom-4">
-      <tui-input-password formControlName="password">Новый пароль</tui-input-password>
-      <tui-field-error formControlName="password"></tui-field-error>
-    </div>
-    <div class="tui-space_bottom-4">
-      <tui-input-password formControlName="confirmPassword">Повторите пароль</tui-input-password>
-      <tui-field-error formControlName="confirmPassword"></tui-field-error>
-    </div>
-  </div>
-</div>
