+++ src/app/modules/profile/_modals/edit-personal-info/edit-personal-info.component.ts
@@ -4,10 +4,16 @@ import {
   ChangeDetectorRef,
   Inject,
   OnInit,
-  Injector,
 } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { POLYMORPHEUS_CONTEXT, PolymorpheusComponent } from '@tinkoff/ng-polymorpheus';
+import {
+  AbstractControl,
+  FormBuilder,
+  FormGroup,
+  ValidationErrors,
+  ValidatorFn,
+  Validators,
+} from '@angular/forms';
+import { POLYMORPHEUS_CONTEXT } from '@tinkoff/ng-polymorpheus';
 import { TuiDestroyService, TuiDialog, TuiStringHandler } from '@taiga-ui/cdk';
 import { TUI_VALIDATION_ERRORS } from '@taiga-ui/kit';
 import { TuiNotification, TuiNotificationsService, TuiValueContentContext } from '@taiga-ui/core';
@@ -17,9 +23,6 @@ import { ModalOptions } from '../../../../components/modal/modal.interface';
 import { isFormInvalid } from '../../../../_utils/formValidCheck';
 import { IdNameResponse } from '../../../../_models/responses/id-name.response';
 import { User } from '../../../../_entities/user.entity';
-import { baseModalButton } from '../../../../_utils/constants';
-import { ModalService } from '../../../../components/modal/modal.service';
-import { ChangePasswordComponent } from './change-password/change-password.component';
 
 @Component({
   selector: 'efes-edit-personal-info',
@@ -39,20 +42,20 @@ import { ChangePasswordComponent } from './change-password/change-password.compo
 })
 export class EditPersonalInfoComponent implements OnInit {
   public userForm: FormGroup;
+  public passwordForm: FormGroup;
   public roles: IdNameResponse[] = [];
+  public passwordChanges: boolean = false;
   private readonly user: User | undefined;
   public readonly content: TuiStringHandler<TuiValueContentContext<number>> = ({ $implicit: id }) =>
     this.getRoleById(id) ? `${this.getRoleById(id)?.name}` : '';
   constructor(
     private fb: FormBuilder,
     private userService: UserService,
-    private modalService: ModalService,
     public cdr: ChangeDetectorRef,
     private destroy$: TuiDestroyService,
     @Inject(POLYMORPHEUS_CONTEXT) public readonly context: TuiDialog<ModalOptions, boolean>,
     @Inject(TuiNotificationsService)
     private readonly notificationsService: TuiNotificationsService,
-    @Inject(Injector) private readonly injector: Injector,
   ) {
     const saveButton = this.context.buttons ? this.context.buttons[1] : undefined;
     const passwordButton = this.context.buttons ? this.context.buttons[0] : undefined;
@@ -74,6 +77,13 @@ export class EditPersonalInfoComponent implements OnInit {
     });
     this.user = this.context.data;
     if (this.user) this.userForm.patchValue(this.user);
+    this.passwordForm = this.fb.group(
+      {
+        password: ['', [Validators.required]],
+        confirmPassword: [''],
+      },
+      { validators: this.checkPasswords },
+    );
   }
 
   ngOnInit() {
@@ -105,19 +115,49 @@ export class EditPersonalInfoComponent implements OnInit {
     );
   }
 
+  public onSubmit(): void {
+    if (isFormInvalid(this.passwordForm, this.cdr)) return;
+    const password = this.passwordForm.controls.password.value;
+    this.userService.setNewPassword(password).subscribe((response) => {
+      console.log(response);
+    });
+  }
+
   public changePassword(): void {
-    this.modalService
-      .open<boolean>(new PolymorpheusComponent(ChangePasswordComponent), {
-        heading: 'Смена пароля',
-        buttons: baseModalButton,
-      })
-      .pipe(takeUntil(this.destroy$))
-      .subscribe((result) => {
-        console.log(result);
-      });
+    if (this.passwordChanges) {
+      this.context.heading = 'Персональная информация';
+      const saveButton = this.context.buttons ? this.context.buttons[1] : undefined;
+      const passwordButton = this.context.buttons ? this.context.buttons[0] : undefined;
+      if (saveButton) {
+        saveButton.value = () => this.onSave();
+        saveButton.voidType = true;
+      }
+      if (passwordButton) {
+        passwordButton.label = 'Сменить пароль';
+      }
+    } else {
+      this.context.heading = 'Смена пароля';
+      const saveButton = this.context.buttons ? this.context.buttons[1] : undefined;
+      const passwordButton = this.context.buttons ? this.context.buttons[0] : undefined;
+      if (saveButton) {
+        saveButton.value = () => this.onSubmit();
+        saveButton.voidType = true;
+      }
+      if (passwordButton) {
+        passwordButton.label = 'К пользователю';
+      }
+    }
+    this.passwordChanges = !this.passwordChanges;
+    this.cdr.detectChanges();
   }
 
   private getRoleById(id: number): IdNameResponse | undefined {
     return this.roles.find((role) => role.id === id);
   }
+
+  private checkPasswords: ValidatorFn = (group: AbstractControl): ValidationErrors | null => {
+    const pass = group.get('password')?.value;
+    const confirmPass = group.get('confirmPassword')?.value;
+    return pass === confirmPass ? null : { notSame: true };
+  };
 }
